pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pygcp"
version = "1.0.2"
authors = [
    { name = "Chan Yan Zhe", email = "yanzhechan29@gmail.com" }
]
description = "A Python library for Google Cloud Platform (GCP) and Google Workspace related operations"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas",
    "google-cloud-bigquery",
    "google-cloud-storage",
    "google-api-python-client",
    "google-auth",
    "openpyxl>=3.0.0",
    "protobuf>=3.20.0,<4.0.0"
]

[project.urls]
Homepage = "https://github.com/nacht29/pygcp"

[tool.setuptools]
# Critical configuration for src-layout:
package-dir = {"" = "src"}
packages = ["pygcp"]

# Optional (recommended) - include package data like non-PY files:
include-package-data = true

[tool.setuptools.package-data]
"pygcp" = ["*.json", "*.yaml", "*.txt"]  # Add other file types if needed